# lib/api-zod/src/generated/types/checkinInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NYCELO API — NYC neighborhood ELO rankings
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinInput {
  voterToken: string;
  lat: number;
  lng: number;
}